+++ app/agora/api-backend-nest/src/accesos/permisos/permisos.service.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@nestjs/common';
+import { PrismaService } from '../../database/prisma/prisma.service';
+
+@Injectable()
+export class PermisosService {
+  constructor(private readonly prisma: PrismaService) {}
+
+  obtenerPermisos() {
+    return this.prisma.permiso.findMany();
+  }
+}