+++ app/agora/api-backend-nest/src/accesos/service-token/service-token.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { ServiceTokenController } from './service-token.controller';
+import { ServiceTokenService } from './service-token.service';
+import { AuthModule } from '../../auth/auth.module';
+
+@Module({
+  imports: [AuthModule],
+  controllers: [ServiceTokenController],
+  providers: [ServiceTokenService],
+})
+export class ServiceTokenModule {}